Everything is TS, another folder reformat, finally did proper libs with d.ts and everything, removed sourcemaps because, let's face it, everything is opensource...why do you need a sourcemap? :P  Added gulp task to build just the libs, and the main build task no longer builds the libs.  Libs emit d.ts, but main sources do not.
Note that this is a potentially breaking update because the SearchFilter now requires my Config lib...which will get pushed with this commit.  My only options were to inline the source or use the published version.  I went with the published version, which means I have to wait for the CDN to pick it up before this version is usable.  Due to that, I have not bumped the version on the user scripts until I can test and makes sure things still work and, if not, fix them...then I'll bump it.